place-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
